requirements.txt
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pydantic>=2.0.0

